+++ apps/services/report.py
@@ -1,7 +1,6 @@
 
 
 from apps.interface.dados_json_caso import Dados_json_caso
-from apps.report.estruturas import Estruturas
 from apps.indicadores.eco_indicadores import EcoIndicadores
 from apps.model.argumento import Argumento
 from inewave.newave import Pmo
@@ -9,6 +8,7 @@ from inewave.newave import Dger
 from inewave.newave import Cvar
 from idecomp.decomp import Relato
 from idessem.dessem.des_log_relato import DesLogRelato
+from apps.report.info import Info
 import plotly.graph_objects as go
 import plotly.io as pio
 import pandas as pd
@@ -22,9 +22,8 @@ import subprocess
 import shutil
 import json
 
-class Report(Estruturas):
+class Report():
     def __init__(self,outpath, arq_json, txt, titulo):
-        Estruturas.__init__(self)
         #self.outpath = outpath
         self.json = arq_json
         self.txt = txt
@@ -103,7 +102,8 @@ class Report(Estruturas):
                         html_file.write(line.strip()+"\n")
 
                     elif("\info") in line:
-                        self.preencheInfo(html_file, data)
+                        info = Info(html_file, data)
+                        html_file.write(info.text_html+"\n")
 
                     elif("\page{") in line:
                         if(flag_primeira_pagina == False):
@@ -177,197 +177,4 @@ class Report(Estruturas):
                 html_file.write(conteudo)
 
         print("Report saved as report.html")
-
-    def preencheInfo(self, html_file, data):
-        Inicio_tabela = """
-<table>
-<tr>
-<th>Nome do Caso</th>
-<th>Caminho</th>
-<th>Modelo</th>
-<th>Cor</th>
-</tr>
-"""                         
-        Template_tabela_caso = """
-<tr>
-<td>nome</td>
-<td>caminho</td>
-<td>modelo</td>
-<td style="background-color: cor;"></td>
-</tr>
-"""                         
-        html_file.write("<h3>Informações Gerais do Estudo</h3>"+"\n")
-        html_file.write(Inicio_tabela)
-        
-        for caso in data.casos:
-            temp = Template_tabela_caso
-            temp = temp.replace("nome", caso.nome)
-            temp = temp.replace("caminho", caso.caminho)
-            temp = temp.replace("modelo", caso.modelo)
-            temp = temp.replace("cor", caso.cor)
-            html_file.write(temp)
-        html_file.write("</table>"+"\n")
-        
-        html_file.write("<h2>Eco Dados Entrada</h2>"+"\n")
-        
-        flag_nw = flag_deco = flag_dss = True
-        for caso in data.casos:
-            if(caso.modelo == "NEWAVE"):
-                if(flag_nw == True):
-                    html_file.write(self.mapa_tabela_modelo[caso.modelo])
-                    flag_nw = False
-                temp = self.preenche_modelo_tabela_modelo_NEWAVE(caso)
-            if(caso.modelo == "DECOMP"):
-                if(flag_deco == True):
-                    html_file.write(self.mapa_tabela_modelo[caso.modelo])
-                    flag_deco = False
-                temp = self.preenche_modelo_tabela_modelo_DECOMP(caso)
-            if(caso.modelo == "DESSEM"):
-                if(flag_dss == True):
-                    html_file.write(self.mapa_tabela_modelo[caso.modelo])
-                    flag_dss = False
-                temp = self.preenche_modelo_tabela_modelo_DESSEM(caso)
-            html_file.write(temp)
-        html_file.write("</table>"+"\n")
-
-
-        #html_file.write("<h2>Operacao</h2>"+"\n")
-        #
-        #flag_nw = flag_deco = flag_dss = True
-        #for caso in data.casos:
-        #    if(caso.modelo == "NEWAVE"):
-        #        if(flag_nw == True):
-        #            html_file.write(self.mapa_tabela_modelo[caso.modelo])
-        #            flag_nw = False
-        #        temp = self.preenche_operacao_NEWAVE(caso)
-        #    if(caso.modelo == "DECOMP"):
-        #        if(flag_deco == True):
-        #            html_file.write(self.mapa_tabela_modelo[caso.modelo])
-        #            flag_deco = False
-        #        temp = self.preenche_operacao_DECOMP(caso)
-        #    if(caso.modelo == "DESSEM"):
-        #        if(flag_dss == True):
-        #            html_file.write(self.mapa_tabela_modelo[caso.modelo])
-        #            flag_dss = False
-        #        temp = self.preenche_operacao_DESSEM(caso)
-        #    html_file.write(temp)
-        #html_file.write("</table>"+"\n")
-
-
-
-    def preenche_operacao_NEWAVE(self,caso):
-
-        #df_temp = self.eco_indicadores.retorna_df_concatenado("TEMPO")
-        #temp = self.mapa_template_tabela_modelo[caso.modelo]
-        #temp = temp.replace("Caso", caso.nome)
-        #temp = temp.replace("Modelo", caso.modelo)
-        #data_pmo = Pmo.read(caso.caminho+"/pmo.dat")
-        #data_dger = Dger.read(caso.caminho+"/dger.dat")
-        #data_cvar = Cvar.read(caso.caminho+"/cvar.dat")
-        #print(data_dger.num_series_sinteticas)
-        #temp = temp.replace("Versao", data_pmo.versao_modelo)
-        #temp = temp.replace("Mes_I", str(data_dger.mes_inicio_estudo))
-        #temp = temp.replace("Ano_I", str(data_dger.ano_inicio_estudo))
-        #temp = temp.replace("Anos_Pos", str(data_dger.num_anos_pos_estudo))
-        #temp = temp.replace("It_Max", str(data_dger.num_max_iteracoes))
-        #temp = temp.replace("It_Min", str(data_dger.num_minimo_iteracoes))
-        #temp = temp.replace("FW", str(data_dger.num_forwards))
-        #temp = temp.replace("BK", str(data_dger.num_aberturas))
-        #temp = temp.replace("N_series_sim_final", str(data_dger.num_series_sinteticas))
-        #tipo_sim_fin = "Ind" if data_dger.agregacao_simulacao_final == 1 else "Agr"
-        #temp = temp.replace("SF_Ind", tipo_sim_fin)
-        #temp = temp.replace("CVAR", str(data_cvar.valores_constantes[0])+"x"+str(data_cvar.valores_constantes[1]))
-        ##df_caso = df_temp.loc[(df_temp["caso"] == caso.nome)]
-        ##tempo_total = df_caso.loc[(df_caso["etapa"] == "Tempo Total")]["tempo"].iloc[0]/60
-        ##iteracoes = data_pmo.convergencia["iteracao"].iloc[-1]
-        ##zinf = data_pmo.convergencia["zinf"].iloc[-1]
-        ##custo_total = data_pmo.custo_operacao_total
-        ##desvio_custo = data_pmo.desvio_custo_operacao_total*1.96
-        ##temp = temp.replace("tempo_total", str(tempo_total))
-        ##temp = temp.replace("iteracoes", str(iteracoes))
-        ##temp = temp.replace("zinf", str(zinf))
-        ##temp = temp.replace("custo_total", str(custo_total))
-        ##temp = temp.replace("desvio_custo", str(desvio_custo))
-#
-        #return temp
-        pass
-
-
-    def preenche_operacao_DECOMP(self,caso):
-        pass
-
-    def preenche_operacao_DESSEM(self,caso):
-        pass
-
-
-    def preenche_modelo_tabela_modelo_NEWAVE(self,caso):
-
-        df_temp = self.eco_indicadores.retorna_df_concatenado("TEMPO")
-
-        tempo_total = iteracoes = zinf = custo_total = desvio_custo = 0
-        versao = "0"
-
-        temp = self.mapa_template_tabela_modelo[caso.modelo]
-        temp = temp.replace("Caso", caso.nome)
-        temp = temp.replace("Modelo", caso.modelo)
-
-        data_pmo = Pmo.read(caso.caminho+"/pmo.dat")
-        data_dger = Dger.read(caso.caminho+"/dger.dat")
-        data_cvar = Cvar.read(caso.caminho+"/cvar.dat")
-
-        temp = temp.replace("Versao", data_pmo.versao_modelo)
-        temp = temp.replace("Mes_I", str(data_dger.mes_inicio_estudo))
-        temp = temp.replace("Ano_I", str(data_dger.ano_inicio_estudo))
-        temp = temp.replace("Anos_Pos", str(data_dger.num_anos_pos_estudo))
-        temp = temp.replace("It_Max", str(data_dger.num_max_iteracoes))
-        temp = temp.replace("It_Min", str(data_dger.num_minimo_iteracoes))
-        temp = temp.replace("FW", str(data_dger.num_forwards))
-        temp = temp.replace("BK", str(data_dger.num_aberturas))
-        temp = temp.replace("N_series_sim_final", str(data_dger.num_series_sinteticas))
-        tipo_sim_fin = "Ind" if data_dger.agregacao_simulacao_final == 1 else "Agr"
-        temp = temp.replace("SF_Ind", tipo_sim_fin)
-        temp = temp.replace("CVAR", str(data_cvar.valores_constantes[0])+"x"+str(data_cvar.valores_constantes[1]))
-        return temp
-
-    def preenche_modelo_tabela_modelo_DECOMP(self, caso):
-
-        df_temp = self.eco_indicadores.retorna_df_concatenado("TEMPO")
-
-        tempo_total = iteracoes = zinf = custo_total = desvio_custo = 0
-        versao = "0"
-
-        temp = self.mapa_template_tabela_modelo[caso.modelo]
-        temp = temp.replace("nome", caso.nome)
-        temp = temp.replace("modelo", caso.modelo)
-        extensao = ""
-        with open(caso.caminho+"/caso.dat") as f:
-            extensao = f.readline().strip('\n')
-        if extensao == "":
-            raise FileNotFoundError(f"Arquivo caso.dat não encontrado.") 
-        data_relato = Relato.read(caso.caminho+"/relato."+extensao).convergencia
-        df_caso = df_temp.loc[(df_temp["caso"] == caso.nome)]
-        tempo_total = df_caso.loc[(df_caso["etapa"] == "Tempo Total")]["tempo"].iloc[0]/60
-        iteracoes = data_relato["iteracao"].iloc[-1]
-        zinf = data_relato["zinf"].iloc[-1]
-        custo_total = " "
-        versao = " "
-        temp = temp.replace("versao", versao)
-        temp = temp.replace("tempo_total", str(tempo_total))
-        temp = temp.replace("iteracoes", str(iteracoes))
-        temp = temp.replace("zinf", str(zinf))
-        temp = temp.replace("custo_total", str(custo_total))
-        return temp
-
-    def preenche_modelo_tabela_modelo_DESSEM(self, caso):
-        df_temp = self.eco_indicadores.retorna_df_concatenado("TEMPO")
-        ## FALTA COMPLEMENTAR, ESTA INCOMPLETO
-        data_relato = DesLogRelato.read(caso.caminho+"/DES_LOG_RELATO.DAT")
-        df_caso = df_temp.loc[(df_temp["caso"] == caso.nome)]
-        tempo_total = df_caso["tempo"].sum()/60
-        iteracoes = " "
-        zinf = " "
-        custo_total = " "
-        desvio_custo = " "
-        versao = data_relato.versao
-        return temp
